third_party/WebKit/LayoutTests/media/mediasession/navigator-mediasession.html

<!DOCTYPE html>
<title>Navigator MediaSession Test</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
  assert_true(window.navigator.mediaSession instanceof MediaSession);
});

</script>
